MaMuT_: disable TrackScheme view for now
At creation, TrackScheme tries to retrieve the thumbnails image from all
spots, which breaks MaMuT in our case, since it uses ImageJ virtual
images (calling the thumbnail generator causes the whole data to be
loaded in memory, which we do not want). So we disable TrackScheme for
now.
Possible solutions are:
- detecting in TrackScheme if the source image is virtual, then skip
  thumnail creation.
- have a thumbnail generator that can operate on virtual images, live.
  This would require the thumnail to be stored somewhere just after spot
  creation or modification, because we do not want to compute them when
  instantiating TrackScheme.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>