MaMuT_: fix GUI main window on windows maschinen
The main GUI window was only showing two buttons on windows, while it
was wroking fine on other OSes.

It turns out this was due to one icon used in JTabbedPane tabs.
On windows, the size of the tab button rwo is set by the largest icon.
In our case, it was 256 pixels....
Fixed it by using icons of 32x32 size.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>