Fix the numerous TrackScheme-instances-on-loading bug.
The GUImodel keeps a reference to each view that is launched in MaMuT,
then save if in the xml file. And restores then upon loading.
Yes.
The problem is that we do not deregister the views when we close a view.
This means that when we re-open the xml file, we generate as many views
as were open from beginning the annotation.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>