Rework the cell division time analyzer.
Made it truly a cell division time analyzer:

It decomposes each track into convex branches and computes the mean and
std of the duration of these branches. Careful: branches that do not
start from a cell division or do not end with a cell division are
excluded from the statistics, so that they do not bias the mean.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>