Import Base64 class from Robert Harder in the project.
Quoted:

I am placing this code in the Public Domain. Do with it as you will.
This software comes with no guarantees or warranties but with plenty of
well-wishing instead! Please visit http://iharder.net/base64
periodically to check for updates or to contribute improvements.

Version:
2.3.7
Author:
Robert Harder
rob@iharder.net