Release v0.19.0
* Like for TrackMate: New display mode: show only selection.

This track display mode only shows the content of the current selection, for
spots and edges.  It can be accessed in the GUI panel 'Configure views' with a
special track display mode called 'Selection only'.

Careful: the content of the selection model can still be edited in this mode.
Of course, the track display mode is ignored by TrackScheme, which then can be
used to modify the selection in this mode.

* Like for TrackMate: The manual editing of spot radius (press 'Q' and 'E') is
fully reversible (if you press 3 times 'Q' then 3 times 'E', you get back the
initial radius).