Fix bookmark creation and navigation bug.
Before this commit, when the user types a letter for the bookmark, the key
pressed event is still caught by the actions. So if you pick the 'A' key, you
will be adding a spot everytime you access this bookmark.

Following advice from Tobias, I made a new BookmarkEditor that correctly
blocks the MaMuT actions.

I could not extend the existing one because currently the static enum Mode
is not visible.