Release v0.21.0.
- Record movies from MaMuT viewer windows.
You can now record movies and max. projection movies from the MaMuT
viewer. The annotations will be recorded as well.

- Use the BDV bookmarks system in MaMuT.
With the same kep mapping by default (Shift-B + some key to store a
bookmark, B + bookmark key to recall it, O + bookmar key to recall just
its orientation).

- Use the BDV grouping system in MaMuT.

- Import bookmarks from existing settings file.
When launching a new MaMuT annotation, if a settings file is found, we
import bookmarks from it. They are then saved inside the MaMuT file.

- Load or determine display defaults for BDV sources:
If a BigDataViewer settings file is found, we fetch the SetupAssignments
settings from it. We cannot fetch the grouping and visibility settings
and the viewer settings, because these settings are proper to a single
viewer. If a settings file is not present, we compute a default brightness
for each source.

- The help window has been recamped to acknowledge these changes.