Properly annotate CellDivisionTimeAnalyzerSpotFactory.
The @Plugin( type = SpotAnalyzerFactory.class ) annotation was missing.
This caused a problem when TrackMate was used in a Fiji installation
were MaMuT is checked out:

The cell division track analyzer provided by MaMuT also works with a dummy
spot feature that copies this track feature value to all the spots of the
branch currently calculated.

This dummy spot analyzer fletches the spot feature map, but because it
was not annotated, it was not discovered by TrackMate, and TrackMate then
did not know about the feature isInt, dimension, etc...

This in turn caused a problem upon saving the TrackMate session: the spot
reported a value for a certain feature, but the TmXmlWriter generated a
NPE trying to access the isInt property of this feature.

Annotating the dummy analyzer fixes the problem.